Trying of 3
Getting an error bc "fcntl" is getting called an it shouldnt be. (Also, I know for a fact I do not have to interposition it).
So have a array of files in .cc that stores the fd, filename, and balance of the files.
In open: not supposed to actually open the files. I use a big if to match up which file is opened - open it to read its initial balance - and then I close it. I put the filname, the index in the array (which gets manipulated to an fd thats not used (i think) in openFile() func in part2_hash.cc, and the balance in the correct entry in the file array.
(inopen and inclose are variables used to prevent infinite recursion)

In the read- fake the read - dont actually read or do anything

in write - fake the write- dont actually write anything - but take what was going to be written to update the balance of the file that is being written to.

in close: where the actual updates happen - open, update file with the balance variable in part2_hash.cc, and close file